fix(instrumentation): consolidare i due register() in un solo file (REVIEW.md #29)
Esistevano due file instrumentation con register() separati: /instrumentation.ts
(keep-alive Supabase + Sentry + onRequestError) e src/instrumentation.ts
(assertIdentityEnv R24 + sentry + migrazioni + backfill trial_vat_ledger).
Next.js 16 invoca un solo hook, quindi uno dei due era codice morto a runtime:
se l'hook attivo era il root, il fail-fast env d'identita' (regola 24) e il
backfill non giravano in produzione.

Fuso tutto in src/instrumentation.ts (survivor citato da regola 24 e skill
db-migrations) ed eliminato il file root. Ordine in register() nodejs:
assertIdentityEnv -> sentry.server.config -> migrazioni -> backfill ->
keep-alive. Aggiunto l'export onRequestError mancante nel file src.

Le migrazioni in prod continuano a girare anche via migrate.js (Dockerfile,
processo separato): nessuna regressione di deploy.

Test: spostato instrumentation-keep-alive.test.ts sotto src/ con path corretti;
src/instrumentation.test.ts ora mocka setInterval/supabase admin e verifica che
il keep-alive parta dopo il backfill (nodejs) e non su edge. Chiuso #29 in
REVIEW.md (punto 1 gia' risolto, punto 2 implementato qui).
