+++ src/instrumentation-keep-alive.test.ts
@@ -1,6 +1,6 @@
 // @vitest-environment node
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
-import { KEEP_ALIVE_INTERVAL_MS } from "../../instrumentation";
+import { KEEP_ALIVE_INTERVAL_MS } from "./instrumentation";
 
 const { mockCreateAdminSupabaseClient, mockLoggerInfo, mockLoggerWarn } =
   vi.hoisted(() => ({
@@ -21,9 +21,6 @@ vi.mock("@sentry/nextjs", () => ({
   captureRequestError: vi.fn(),
 }));
 
-vi.mock("../../sentry.server.config", () => ({}));
-vi.mock("../../sentry.edge.config", () => ({}));
-
 describe("KEEP_ALIVE_INTERVAL_MS", () => {
   it("vale esattamente 5 giorni in millisecondi", () => {
     expect(KEEP_ALIVE_INTERVAL_MS).toBe(5 * 24 * 60 * 60 * 1000);
@@ -42,7 +39,7 @@ describe("startSupabaseKeepAlive()", () => {
     // Reset dello stato module-level (guardia di idempotenza) tra i test:
     // senza questo, dal secondo test in poi la guardia farebbe early-return.
     vi.resetModules();
-    ({ startSupabaseKeepAlive } = await import("../../instrumentation"));
+    ({ startSupabaseKeepAlive } = await import("./instrumentation"));
 
     mockUnref = vi.fn();
     const mockTimer = { unref: mockUnref } as unknown as ReturnType<
@@ -128,63 +125,6 @@ describe("startSupabaseKeepAlive()", () => {
   it("due chiamate consecutive avviano un solo setInterval (idempotenza)", () => {
     startSupabaseKeepAlive();
     startSupabaseKeepAlive();
-
-    expect(global.setInterval).toHaveBeenCalledOnce();
-  });
-});
-
-describe("register()", () => {
-  let originalNextRuntime: string | undefined;
-  let mockUnref: ReturnType<typeof vi.fn>;
-  let register: () => Promise<void>;
-
-  beforeEach(async () => {
-    // Reset dello stato module-level (guardia di idempotenza) tra i test.
-    vi.resetModules();
-    ({ register } = await import("../../instrumentation"));
-
-    originalNextRuntime = process.env.NEXT_RUNTIME;
-    mockUnref = vi.fn();
-    const mockTimer = { unref: mockUnref } as unknown as ReturnType<
-      typeof setInterval
-    >;
-    vi.spyOn(global, "setInterval").mockReturnValue(mockTimer);
-  });
-
-  afterEach(() => {
-    if (originalNextRuntime === undefined) {
-      delete process.env.NEXT_RUNTIME;
-    } else {
-      process.env.NEXT_RUNTIME = originalNextRuntime;
-    }
-    vi.restoreAllMocks();
-    vi.clearAllMocks();
-  });
-
-  it("avvia il keep-alive quando NEXT_RUNTIME=nodejs", async () => {
-    process.env.NEXT_RUNTIME = "nodejs";
-
-    await register();
-
-    expect(global.setInterval).toHaveBeenCalledWith(
-      expect.any(Function),
-      KEEP_ALIVE_INTERVAL_MS,
-    );
-  });
-
-  it("non avvia il keep-alive quando NEXT_RUNTIME=edge", async () => {
-    process.env.NEXT_RUNTIME = "edge";
-
-    await register();
-
-    expect(global.setInterval).not.toHaveBeenCalled();
-  });
-
-  it("due register() consecutive (nodejs) avviano un solo setInterval (idempotenza)", async () => {
-    process.env.NEXT_RUNTIME = "nodejs";
-
-    await register();
-    await register();
 
     expect(global.setInterval).toHaveBeenCalledOnce();
   });
