+++ src/instrumentation.ts
@@ -1,3 +1,37 @@
+import * as Sentry from "@sentry/nextjs";
+
+// Supabase free tier pausa i progetti dopo 7 giorni senza query al DB.
+// Questo interval esegue una query lightweight ogni 5 giorni per prevenire la pausa.
+// TODO: rimuovere quando si passa a Supabase Pro.
+export const KEEP_ALIVE_INTERVAL_MS = 5 * 24 * 60 * 60 * 1000; // 5 giorni
+
+// Guardia di idempotenza: register() può essere invocata più volte per lo stesso
+// deploy (osservati due ping reali a ~13s di distanza in prod, REVIEW.md #29).
+// Senza questa guardia ogni invocazione impilerebbe un nuovo setInterval.
+let keepAliveStarted = false;
+
+export function startSupabaseKeepAlive() {
+  if (keepAliveStarted) return;
+  keepAliveStarted = true;
+
+  const interval: ReturnType<typeof setInterval> = setInterval(async () => {
+    try {
+      const { createAdminSupabaseClient } =
+        await import("@/lib/supabase/admin");
+      const supabase = createAdminSupabaseClient();
+      await supabase.from("profiles").select("id").limit(1);
+      const { logger } = await import("@/lib/logger");
+      logger.info("Supabase keep-alive ping eseguito");
+    } catch (err) {
+      const { logger } = await import("@/lib/logger");
+      logger.warn({ err }, "Supabase keep-alive ping fallito");
+    }
+  }, KEEP_ALIVE_INTERVAL_MS);
+
+  // .unref() evita che l'interval blocchi lo shutdown del processo Node.js
+  interval.unref();
+}
+
 export async function register() {
   if (process.env.NEXT_RUNTIME === "nodejs") {
     // Fail-fast sulle env d'identita' (NEXT_PUBLIC_APP_URL, *_HOSTNAME, …).
@@ -56,9 +90,16 @@ export async function register() {
     }
 
     await client.end();
+
+    // Keep-alive Supabase in coda al boot nodejs (dopo migrazioni e backfill):
+    // la guardia di idempotenza dentro startSupabaseKeepAlive() evita timer
+    // duplicati su invocazioni multiple di register().
+    startSupabaseKeepAlive();
   }
 
   if (process.env.NEXT_RUNTIME === "edge") {
     await import("../sentry.edge.config");
   }
 }
+
+export const onRequestError = Sentry.captureRequestError;
